refactor(stats): extract the 6 inline StatsSection cards into components
Tier 3 architecture review (A2a).

6 of the ~11 stats cards were hand-written glass-card JSX inline in
StatsSection (with duplicated header markup), while the rest were already
extracted components. Moved them out verbatim: AuthorsCard,
AwardsProgressCard, YearlyCard, OwnedUnreadCard, LibraryProgressCard,
IdentifiedLibraryCard. Prop types are derived via React.ComponentProps
from the item components, so they stay in sync.

StatsSection drops 372 → 244 lines and all cards are now uniform. No
behavior change. Suite 474 green, lint clean, build OK.

(A2b — lifting the DnD masonry into a hook — is a separate PR.)
