+++ src/components/StatsSection.tsx
@@ -1,22 +1,21 @@
 import React, { useState, useEffect } from "react";
 import { motion } from "motion/react";
-import { Book, BarChart3, Award, User, Calendar, Package, RefreshCw, Library, Search, GripVertical, LayoutGrid, RotateCcw, Check } from "lucide-react";
+import { BarChart3, RefreshCw, GripVertical, LayoutGrid, RotateCcw, Check } from "lucide-react";
 import { useStats } from "../hooks/useStats";
 import { useLibraryCheck } from "../hooks/useLibraryCheck";
 import { useMarkAsRead } from "../hooks/useMarkAsRead";
-import { ProgressBar } from "./stats/ProgressBar";
-import { YearlyProgressItem } from "./stats/YearlyProgressItem";
-import { LibraryProgressItem } from "./stats/LibraryProgressItem";
-import { AuthorProgressItem } from "./stats/AuthorProgressItem";
-import { IdentifiedLibraryItem } from "./stats/IdentifiedLibraryItem";
-import { OwnedUnreadItem } from "./stats/OwnedUnreadItem";
-import { AwardCoverageGrid } from "./stats/AwardCoverageGrid";
 import { AvailabilityCard } from "./stats/AvailabilityCard";
 import { PublishingCard } from "./stats/PublishingCard";
 import { DecadeHistogram } from "./stats/DecadeHistogram";
 import { MarketCard } from "./stats/MarketCard";
 import { CyclesHarvestCard } from "./stats/CyclesHarvestCard";
 import { KpiRow } from "./stats/KpiRow";
+import { AuthorsCard } from "./stats/AuthorsCard";
+import { AwardsProgressCard } from "./stats/AwardsProgressCard";
+import { YearlyCard } from "./stats/YearlyCard";
+import { OwnedUnreadCard } from "./stats/OwnedUnreadCard";
+import { LibraryProgressCard } from "./stats/LibraryProgressCard";
+import { IdentifiedLibraryCard } from "./stats/IdentifiedLibraryCard";
 import { useEffectiveConfig, persistStatsOrder } from "../hooks/useAppConfig";
 import { orderByIds, moveId, distributeColumns } from "../utils/statsLayout";
 
@@ -90,143 +89,17 @@ export const StatsSection: React.FC = () => {
   // Card definitions in code order (= default). Each has a stable `id`, by which
   // we remember the user's layout (ui.statsOrder). Reordering doesn't change card content.
   const cards: StatCard[] = [
-    {
-      id: "authors",
-      node: (
-        <motion.div initial={{ opacity: 0, y: 20 }} animate={{ opacity: 1, y: 0 }} className="glass-card p-6 rounded-3xl border-cyan-500/10 space-y-6">
-          <h3 className="text-sm font-bold font-display uppercase tracking-widest text-cyan-500 flex items-center gap-2 mb-4">
-            <User className="w-4 h-4" />
-            Indeks Autorów
-          </h3>
-          <div className="space-y-4 max-h-[300px] overflow-y-auto pr-2 custom-scrollbar">
-            {stats.authorStats.map((author) => (
-              <AuthorProgressItem key={author.name} author={author} />
-            ))}
-          </div>
-        </motion.div>
-      ),
-    },
-    {
-      id: "awards",
-      node: (
-        <motion.div initial={{ opacity: 0, y: 20 }} animate={{ opacity: 1, y: 0 }} transition={{ delay: 0.1 }} className="glass-card p-6 rounded-3xl border-purple-500/10 space-y-6">
-          <h3 className="text-sm font-bold font-display uppercase tracking-widest text-purple-500 flex items-center gap-2 mb-4">
-            <Award className="w-4 h-4" />
-            Postęp kolekcji
-          </h3>
-          <div className="space-y-6">
-            <ProgressBar current={stats.awardBooksStats.read} total={stats.awardBooksStats.total} label="Polskie wydania z listy nagród" icon={Book} color="purple" />
-            <ProgressBar current={stats.allAwardsStats.read} total={stats.allAwardsStats.total} label="Książki mające Wszystkie Nagrody" icon={Award} color="emerald" />
-            <AwardCoverageGrid coverage={stats.awardCoverage} />
-          </div>
-        </motion.div>
-      ),
-    },
+    { id: "authors", node: <AuthorsCard authors={stats.authorStats} /> },
+    { id: "awards", node: <AwardsProgressCard awardBooks={stats.awardBooksStats} allAwards={stats.allAwardsStats} coverage={stats.awardCoverage} /> },
     { id: "availability", node: <AvailabilityCard stats={stats.availabilityStats} /> },
     { id: "market", node: <MarketCard market={stats.marketStats} /> },
     { id: "publishing", node: <PublishingCard publishers={stats.publisherStats} series={stats.seriesStats} cycles={stats.cycleStats} /> },
     { id: "cyclesHarvest", node: <CyclesHarvestCard refreshSignal={refreshTick} /> },
     { id: "decades", span2: true, node: <DecadeHistogram decades={stats.decadeStats} /> },
-    {
-      id: "yearly",
-      node: (
-        <motion.div initial={{ opacity: 0, y: 20 }} animate={{ opacity: 1, y: 0 }} transition={{ delay: 0.2 }} className="glass-card p-6 rounded-3xl border-orange-500/10 space-y-6">
-          <h3 className="text-sm font-bold font-display uppercase tracking-widest text-orange-500 flex items-center gap-2 mb-4">
-            <Calendar className="w-4 h-4" />
-            Chronologia Przeczytanych
-          </h3>
-          <div className="space-y-4 max-h-[400px] overflow-y-auto pr-2 custom-scrollbar">
-            {stats.yearlyStats.map((year) => (
-              <YearlyProgressItem key={year.year} year={year} />
-            ))}
-          </div>
-        </motion.div>
-      ),
-    },
-    {
-      id: "ownedUnread",
-      node: (
-        <motion.div initial={{ opacity: 0, y: 20 }} animate={{ opacity: 1, y: 0 }} transition={{ delay: 0.25 }} className="glass-card p-6 rounded-3xl border-emerald-500/10 space-y-6">
-          <h3 className="text-sm font-bold font-display uppercase tracking-widest text-emerald-500 flex items-center gap-2 mb-4">
-            <Package className="w-4 h-4" />
-            Zasoby Oczekujące (Posiadane)
-          </h3>
-          <div className="space-y-3 max-h-[300px] overflow-y-auto pr-2 custom-scrollbar">
-            {stats.ownedUnread.length === 0 ? (
-              <p className="text-slate-400 text-sm italic text-center py-8">Wszystkie posiadane zasoby zostały przyswojone.</p>
-            ) : (
-              stats.ownedUnread.map((book, idx) => (
-                <OwnedUnreadItem key={idx} book={book} marking={markingId === book.id} disabled={markingId !== null} onMarkAsRead={markAsRead} />
-              ))
-            )}
-          </div>
-        </motion.div>
-      ),
-    },
-    {
-      id: "library",
-      node: (
-        <motion.div initial={{ opacity: 0, y: 20 }} animate={{ opacity: 1, y: 0 }} transition={{ delay: 0.3 }} className="glass-card p-6 rounded-3xl border-blue-500/10 space-y-6">
-          <h3 className="text-sm font-bold font-display uppercase tracking-widest text-blue-500 flex items-center gap-2 mb-4">
-            <Library className="w-4 h-4" />
-            Książki dostępne w bibliotekach
-          </h3>
-          <div className="space-y-4 max-h-[400px] overflow-y-auto pr-2 custom-scrollbar">
-            {stats.libraryStats.map((library) => (
-              <LibraryProgressItem key={library.id} library={library} onMarkAsRead={markAsRead} markingId={markingId} />
-            ))}
-          </div>
-        </motion.div>
-      ),
-    },
-    {
-      id: "identified",
-      node: (
-        <motion.div initial={{ opacity: 0, y: 20 }} animate={{ opacity: 1, y: 0 }} transition={{ delay: 0.35 }} className="glass-card p-6 rounded-3xl border-blue-500/10 space-y-6">
-          <div className="flex items-center justify-between mb-4">
-            <h3 className="text-sm font-bold font-display uppercase tracking-widest text-blue-400 flex items-center gap-2">
-              <Search className="w-4 h-4" />
-              Zidentyfikowane w bibliotekach
-            </h3>
-            <button
-              onClick={() => checkAllLibraries(branches)}
-              disabled={checkingLibrary !== null}
-              className="px-4 py-2 bg-blue-500/20 hover:bg-blue-500/30 text-blue-300 rounded-xl border border-blue-500/30 text-xs font-bold uppercase tracking-widest transition-all disabled:opacity-50 flex items-center gap-2"
-            >
-              {checkingLibrary ? (
-                <motion.div animate={{ rotate: 360 }} transition={{ repeat: Infinity, duration: 1, ease: "linear" }}>
-                  <RefreshCw className="w-3 h-3" />
-                </motion.div>
-              ) : (
-                <Search className="w-3 h-3" />
-              )}
-              Skanuj Wszystkie
-            </button>
-          </div>
-          {libraryError && (
-            <div className="p-3 bg-red-500/10 border border-red-500/20 rounded-xl text-xs text-red-400 font-bold mb-4">
-              Błąd skanowania: {libraryError}
-            </div>
-          )}
-          <div className="space-y-6 max-h-[500px] overflow-y-auto pr-2 custom-scrollbar">
-            {branches.map((library) => (
-              <IdentifiedLibraryItem
-                key={library.id}
-                library={library}
-                books={identifiedBooks[library.id] || []}
-                onCheck={() => checkLibrary(library.id, library.code)}
-                onStop={stopLibraryCheck}
-                onMarkAsRead={markAsRead}
-                markingId={markingId}
-                markedIds={markedIds}
-                isChecking={checkingLibrary === library.id}
-                progress={checkingLibrary === library.id ? checkProgress : null}
-              />
-            ))}
-          </div>
-        </motion.div>
-      ),
-    },
+    { id: "yearly", node: <YearlyCard years={stats.yearlyStats} /> },
+    { id: "ownedUnread", node: <OwnedUnreadCard books={stats.ownedUnread} markingId={markingId} onMarkAsRead={markAsRead} /> },
+    { id: "library", node: <LibraryProgressCard libraries={stats.libraryStats} onMarkAsRead={markAsRead} markingId={markingId} /> },
+    { id: "identified", node: <IdentifiedLibraryCard branches={branches} identifiedBooks={identifiedBooks} onCheck={checkLibrary} onCheckAll={() => checkAllLibraries(branches)} onStop={stopLibraryCheck} checkingLibrary={checkingLibrary} checkProgress={checkProgress} libraryError={libraryError} onMarkAsRead={markAsRead} markingId={markingId} markedIds={markedIds} /> },
   ];
 
   // Effective order (user's save + new cards at the end) and lookup by id.
